+++ src/Modules/GraphQL/Type/Fields/EnqueuedScript.php
@@ -3,6 +3,7 @@
  * Registers custom fields to the GraphQL EnqueuedScript Object.
  *
  * @todo Temporary (non-conflicting) backport until supported by WPGraphQL.
+ * @todo Remove when WPGraphQL >= 1.30.0 is required.
  *
  * @see https://github.com/wp-graphql/wp-graphql/pull/3196
  *
@@ -25,7 +26,7 @@ final class EnqueuedScript extends AbstractFields {
 	}
 
 	/**
-	 * Register fields to the Type.
+	 * {@inheritDoc}
 	 */
 	public function register(): void {
 		// Early return if WPGraphQL will register the field itself.